refactor(module:highlight): deprecate `NzHighlightModule`
