+++ components/core/util/encode-entities.ts
@@ -1,0 +1,26 @@
+/**
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
+ */
+
+const SURROGATE_PAIR_REGEXP = /[\uD800-\uDBFF][\uDC00-\uDFFF]/g;
+// ! to ~ is the ASCII range.
+const NON_ALPHANUMERIC_REGEXP = /([^#-~ |!])/g;
+
+/**
+ * Escapes all potentially dangerous characters, so that the
+ * resulting string can be safely inserted into attribute or
+ * element text.
+ */
+export function encodeEntities(value: string): string {
+  return value
+    .replace(/&/g, '&amp;')
+    .replace(SURROGATE_PAIR_REGEXP, (match: string) => {
+      const hi = match.charCodeAt(0);
+      const low = match.charCodeAt(1);
+      return `&#${(hi - 0xd800) * 0x400 + (low - 0xdc00) + 0x10000};`;
+    })
+    .replace(NON_ALPHANUMERIC_REGEXP, (match: string) => `&#${match.charCodeAt(0)};`)
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;');
+}
