+++ components/core/util/public-api.ts
@@ -24,3 +24,4 @@ export * from './dynamic-css';
 export * from './status-util';
 export * from './from-event-outside-angular';
 export * from './variant-utils';
+export * from './encode-entities';
